analytics/ingest/requirements.txt
google-api-python-client
google-auth-oauthlib
google-cloud-bigquery
beautifulsoup4
html5lib
tldextract
requests
python-dateutil
